Added cards for USS Cairo
We need some Java Scripting for delta_shift_cairo to make it so it can be equipped without taking up a slot. I suggest that it be made to just duplicate whatever slot it is placed in. Example if I put Delta Shift in a Tech slot another Tech slot is created for the ship it is placed on.