New idea
Seeing as useing the name will spit out every card sharing that name I feel I need to cut that down. So instead of useing the name to name I will try to go with also id to description. This may take a few atempts in both Utopia and TTS